pyproject.toml
[project]
name = "mcp-redmine"
version = "2026.01.01.000001"
description = "A MCP server that connects to your Redmine instance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "openapi-core>=0.19.4",
    "pyyaml>=6.0.2",
    "starlette>=0.30.0",
    "uvicorn>=0.30.0",
]
authors = [
  { name="Rune Kaagaard" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license-files = ["LICENSE"]
packages = [
    {include = "mcp_redmine"}
]

[project.scripts]
mcp-redmine = "mcp_redmine.server:main"

[project.urls]
Homepage = "https://github.com/runekaagaard/mcp-redmine"
Issues = "https://github.com/runekaagaard/mcp-redmine/issues"
Disussions = "https://github.com/runekaagaard/mcp-redmine/discussions"

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "hatchling>=1.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=mcp_redmine",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
